requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
appdirs==1.4.4
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
asyncer==0.0.2
attrs==22.2.0
backcall==0.2.0
bcrypt==4.0.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
executing==1.2.0
fastapi==0.87.0
filetype==1.2.0
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.0.3
flatbuffers==23.1.21
fonttools==4.38.0
frozenlist==1.3.3
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.51.1
h11==0.14.0
h5py==3.8.0
humanfriendly==10.0
idna==3.4
ImageHash==4.3.1
imageio==2.26.0
ipykernel==6.21.2
ipython==8.10.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.0.3
jupyter_core==5.2.0
jyserver==0.0.5
keras==2.11.0
kiwisolver==1.4.4
libclang==15.0.6.1
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
mpmath==1.2.1
multidict==6.0.4
nest-asyncio==1.5.6
networkx==3.0
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
onnxruntime==1.13.1
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
packaging==23.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.0.0
pooch==1.6.0
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.5
Pygments==2.14.0
PyMatting==1.1.8
pyparsing==3.0.9
pyreadline3==3.4.1
python-dateutil==2.8.2
python-multipart==0.0.6
PyWavelets==1.4.1
pywin32==305
pyzmq==25.0.0
rembg==2.0.30
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scipy==1.9.3
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.4
stack-data==0.6.2
starlette==0.21.0
sympy==1.11.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow @ file:///C:/Users/Sony/Documents/Set%20ups%20installed/tensorflow-2.11.0-cp310-cp310-win_amd64.whl
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
tifffile==2023.2.28
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
watchdog==2.1.9
wcwidth==0.2.6
Werkzeug==2.2.2
wrapt==1.14.1
yarl==1.8.2
